Arduino Code for Robot
This code was built using Arduino IDE version 1.8.13. Some additional libraries may need to be installed. Libraries are listed in the beginning and can be downloaded using Library manager found in the "Tools" tab of Arduino IDE. Messages are printed to Serial monitor as well LCD screen so user can see what part of the code is being executed. ESP01 is setup manually.